# docs/openapi/sak-admin.v0.yaml
...
openapi: 3.0.3
info:
  title: SwissArmyNoife HTTP Admin
  version: 0.1.0
  description: |
    Hand-maintained OpenAPI v0 stub (`sak323-d`). Not codegen output — types/OpenAPI generation
    from `types` is deferred to a future epic. Broker admin routes (`sak323-a`).
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      summary: OpenAI-shaped chat facade over a pre-bound llm.chat (`sak540-b`)
      description: >
        Requires `binding_id` (body or `X-Sak-Binding-Id`). `stream=true` is rejected in v0.
        Tools round-trip is sak540-c.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                binding_id:
                  type: string
                messages:
                  type: array
                  items:
                    type: object
                model:
                  type: string
                stream:
                  type: boolean
      responses:
        "200":
          description: OpenAI-ish chat.completion
        "400":
          description: Invalid request / stream not supported / wrong offer
...